httplib2 is now optional (but strongly recommended).
The httplib2 transport has been punted to an optional module (with a corresponding optional dependency). If httplib2 is unavailable at runtime, refreshbooks now emits a warning about SSL and uses urllib2 instead. (This also limits the sophistication of error handling, as urllib2 makes it hard to obtain the content of an HTTP error response.)